# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateEmailTemplateRequest:
    properties:
      html_body:
        description: HTML email body
        example: <p>Hello {{username}}, your verification code is {{code}}</p>
        type: string
      name:
        description: Template name (max 100 characters)
        example: Email Verification Template
        maxLength: 100
        type: string
      subject:
        description: Email subject line (max 200 characters)
        example: Verify your email address
        maxLength: 200
        type: string
      text_body:
        description: Plain text email body
        example: Hello {{username}}, your verification code is {{code}}
        type: string
      type:
        description: 'Template type: verification, password_reset, welcome, 2fa, or
          custom'
        enum:
        - verification
        - password_reset
        - welcome
        - 2fa
        - custom
        example: verification
        type: string
      variables:
        description: Available variable names for template
        example:
        - username
        - email
        - code
        - expiry_minutes
        items:
          type: string
        type: array
    required:
    - html_body
    - name
    - subject
    - type
    type: object
...
  models.EmailTemplate:
    properties:
      created_at:
        type: string
      html_body:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        type: string
      subject:
        maxLength: 200
        type: string
      text_body:
        type: string
      type:
        enum:
        - verification
        - password_reset
        - welcome
        - 2fa
        type: string
      updated_at:
        type: string
      variables:
        description: Available variables as JSON array
        items:
          type: integer
        type: array
    required:
    - html_body
    - name
    - subject
    - type
    type: object
  models.EmailTemplateListResponse:
    properties:
      page:
        description: Current page number
        example: 1
        type: integer
      per_page:
        description: Number of items per page
        example: 20
        type: integer
      templates:
        description: List of email templates
        items:
          $ref: '#/definitions/models.EmailTemplate'
        type: array
      total:
        description: Total number of templates
        example: 10
        type: integer
      total_pages:
        description: Total number of pages
        example: 1
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      details:
        description: Additional error details
        example: Email field is required
        type: string
      error:
        description: HTTP error status text
        example: Bad Request
        type: string
      message:
        description: Human-readable error message
        example: Invalid request parameters
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        description: Response message
        example: Operation completed successfully
        type: string
    type: object
...
  models.PreviewEmailTemplateRequest:
    properties:
      html_body:
        description: HTML template to preview
        example: <p>Hello {{username}}</p>
        type: string
      text_body:
        description: Text template to preview
        example: Hello {{username}}
        type: string
      variables:
        additionalProperties:
          type: string
        description: Sample variable values for preview
        example:
          email: john@example.com
          username: John Doe
        type: object
    required:
    - html_body
    type: object
...
  models.UpdateEmailTemplateRequest:
    properties:
      html_body:
        description: HTML email body
        example: <p>Updated HTML content</p>
        type: string
      is_active:
        description: Whether the template is active
        example: true
        type: boolean
      name:
        description: Template name (max 100 characters)
        example: Updated Template Name
        maxLength: 100
        type: string
      subject:
        description: Email subject line (max 200 characters)
        example: Updated subject
        maxLength: 200
        type: string
      text_body:
        description: Plain text email body
        example: Updated text content
        type: string
      variables:
        description: Available variable names for template
        example:
        - username
        - email
        items:
          type: string
        type: array
    type: object
...
host: localhost:8811
info:
  contact:
    email: maksbalashov@gmail.com
    name: API Support
  description: |-
    Centralized authentication and authorization system for microservices ecosystem.

    ## Overview
    Auth Gateway provides a complete authentication and authorization solution including:
    - User registration and authentication (email/password, OAuth, passwordless)
    - Two-factor authentication (TOTP)
    - API key management for service-to-service communication
    - Role-based access control (RBAC)
    - Session management
    - Webhooks for event notifications

    ## Authentication
    Most endpoints require authentication via JWT Bearer token or API key.

    ### JWT Authentication
    Include the access token in the Authorization header: `Authorization: Bearer {token}`

    ### API Key Authentication
    Include the API key in the X-API-Key header: `X-API-Key: agw_{key}`
  title: Auth Gateway API
  version: "1.0"
paths:
...
  /api/admin/templates:
    get:
      consumes:
      - application/json
      description: Get a list of all email templates (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailTemplateListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all email templates
      tags:
      - Admin - Email Templates
    post:
      consumes:
      - application/json
      description: Create a new email template for customized emails (admin only)
      parameters:
      - description: Template data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateEmailTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.EmailTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new email template
      tags:
      - Admin - Email Templates
  /api/admin/templates/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an email template by ID (admin only)
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an email template
      tags:
      - Admin - Email Templates
    get:
      consumes:
      - application/json
      description: Get a specific email template by its ID (admin only)
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailTemplate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get email template by ID
      tags:
      - Admin - Email Templates
    put:
      consumes:
      - application/json
      description: Update an existing email template (admin only)
      parameters:
      - description: Template ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Template update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEmailTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an email template
      tags:
      - Admin - Email Templates
  /api/admin/templates/preview:
    post:
      consumes:
      - application/json
      description: Render a template with sample data for preview (admin only)
      parameters:
      - description: Preview data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PreviewEmailTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PreviewEmailTemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Preview an email template
      tags:
      - Admin - Email Templates
  /api/admin/templates/types:
    get:
      consumes:
      - application/json
      description: Get a list of all available email template types (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TemplateTypesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get available template types
      tags:
      - Admin - Email Templates
  /api/admin/templates/variables/{type}:
    get:
      consumes:
      - application/json
      description: Get the list of default variables available for a specific template
        type (admin only)
      parameters:
      - description: Template type (e.g., welcome, password_reset, verification)
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TemplateVariablesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get default variables for a template type
      tags:
      - Admin - Email Templates
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key for service-to-service authentication. Format: "agw_{key}"'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token for user authentication. Format: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"